wip: clean up
